# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 重点项目科研管理系统 API
  version: 1.0.0
  description: |
    当前 React 原型对应的后端契约。ID 在 JSON 中统一使用 string；时间为 RFC 3339，日期为 ISO 8601。
    归档材料无审批流程；成果和月季报采用初审、终审并保留提交版本快照。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: roleId, in: query, schema: { type: string } }
        - { name: enabled, in: query, schema: { type: boolean } }
      responses:
        '200': { description: 用户分页, content: { application/json: { schema: { $ref: '#/components/schemas/UserPage' } } } }
    post:
      tags: [Users]
      operationId: createUser
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserCreateRequest' } } } }
      responses:
        '201': { description: 已使用管理员指定的登录密码创建, content: { application/json: { schema: { $ref: '#/components/schemas/UserCreateResponse' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/{userId}:
    parameters: [{ $ref: '#/components/parameters/UserId' }]
    get:
      tags: [Users]
      operationId: getUser
      responses: { '200': { description: 用户详情, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } } }
    patch:
      tags: [Users]
      operationId: updateUserProfile
      description: 编辑用户名、角色、联系人、手机和邮箱；平台角色与课题单位角色不能直接互换，修改角色后旧会话失效。
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserUpdateRequest' } } } }
      responses: { '200': { description: 已更新, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } } }
  /users/{userId}/status:
    put:
      tags: [Users]
      operationId: setUserStatus
      parameters: [{ $ref: '#/components/parameters/UserId' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [enabled], properties: { enabled: { type: boolean } } } } } }
      responses: { '200': { description: 已更新；停用时现有会话立即失效, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } } }
  '/users/{userId}/password':
    put:
      tags: [Users]
      operationId: changeUserPassword
      parameters: [{ $ref: '#/components/parameters/UserId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordChangeRequest' } } } }
      responses:
        '204': { description: 密码已修改，账号原有会话立即失效 }
...
  /roles/{roleId}:
    get:
      tags: [Users]
      operationId: getRole
      parameters: [{ $ref: '#/components/parameters/RoleId' }]
      responses: { '200': { description: 角色详情, content: { application/json: { schema: { $ref: '#/components/schemas/Role' } } } } }
    put:
      tags: [Users]
      operationId: updateRolePermissions
      parameters: [{ $ref: '#/components/parameters/RoleId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RolePermissionRequest' } } } }
      responses: { '200': { description: 已更新并清除权限缓存, content: { application/json: { schema: { $ref: '#/components/schemas/Role' } } } } }
...
  /units:
    get:
      tags: [Users]
      operationId: listUnits
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
        - { name: internal, in: query, schema: { type: boolean } }
      responses: { '200': { description: 单位列表, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Unit' } } } } } }
...
  /topics/{topicId}:
    parameters: [{ $ref: '#/components/parameters/TopicId' }]
    get:
      tags: [Topics]
      operationId: getTopic
      responses: { '200': { description: 课题详情, content: { application/json: { schema: { $ref: '#/components/schemas/Topic' } } } }, '403': { $ref: '#/components/responses/Forbidden' }, '401': { $ref: '#/components/responses/Unauthorized' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationError' }, '404': { description: '课题或课题内成员不存在', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } } }
    put:
      tags: [Topics]
      operationId: updateTopic
      description: 仅科研助理维护。recordVersion在编辑时必须携带最新值，否则409。participantUnitIds省略或空数组时保留原成员，传入时只补充新关系，已有关系包括停用关系不自动改变。更换牵头时原子更新，新牵头原有关系恢复有效，旧牵头保留为有效承担单位。暂停、关闭或停用时拒绝普通编辑。
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TopicWriteRequest' } } } }
      responses: { '200': { description: 已更新, content: { application/json: { schema: { $ref: '#/components/schemas/Topic' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationError' }, '404': { description: '课题或课题内成员不存在', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } } }
...
  /topics/{topicId}/unit-allocations:
    get:
      tags: [Indicators]
      operationId: listUnitAllocations
      description: 默认view=effective读取生效分配；管理角色和当前牵头看全课题，承担单位只能看自身。view=draft仅当前牵头且有unit-allocation.manage权限可读，返回X-Draft-Version和X-Topic-Indicator-Version响应头。
      parameters: [{ $ref: '#/components/parameters/TopicId' }, { name: nodeId, in: query, required: true, schema: { type: string } }, { name: view, in: query, schema: { type: string, enum: [effective, draft], default: effective } }]
      responses: { '200': { description: 当前节点的本阶段单位分配, headers: { X-Draft-Version: { description: 草稿版本，仅草稿查询返回, schema: { type: integer, minimum: 0 } }, X-Topic-Indicator-Version: { description: 草稿绑定的课题指标发布版本，仅草稿查询返回, schema: { type: integer, minimum: 0 } } }, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/UnitIndicatorAllocation' } } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/ValidationError' } }
    put:
      tags: [Indicators]
      operationId: saveUnitAllocations
      description: 当前有效牵头且有unit-allocation.manage权限。完整替换节点的本阶段分配草稿，首次draftVersion省略或0，后续必须匹配响应X-Draft-Version；每行单位/定义不能重复，单位须为有效且启用成员，目标须已由课题下发。
      parameters: [{ $ref: '#/components/parameters/TopicId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UnitAllocationBatch' } } } }
      responses: { '200': { description: 已保存, headers: { X-Draft-Version: { description: 保存后的草稿版本, schema: { type: integer, minimum: 1 } }, X-Topic-Indicator-Version: { description: 草稿绑定的课题指标发布版本, schema: { type: integer, minimum: 1 } } }, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/UnitIndicatorAllocation' } } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationError' } }
  '/topics/{topicId}/unit-allocations:publish':
    post:
      tags: [Indicators]
      operationId: publishUnitAllocations
      description: 当前有效牵头且有unit-allocation.publish权限。必须覆盖所有有效且启用成员（含牵头）和该节点全部生效指标，零目标显式填0；每项单位合计必须等于课题本阶段要求。课题目标版本变化后须重新保存草稿。请求绑定draftVersion；同一用户同键同请求重放204，不同请求409。
      parameters: [{ $ref: '#/components/parameters/TopicId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [nodeId, draftVersion], properties: { nodeId: { type: string }, draftVersion: { type: integer, minimum: 1, maximum: 2147483647 } } } } } }
      responses: { '204': { description: 已下发或成功重放 }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationError' } }
  '/topics/{topicId}/unit-allocations:confirm':
    put:
      tags: [Indicators]
      operationId: confirmUnitAllocations
      description: 当前有效牵头一次确认单位分配并立即生效；草稿写入与发布处于同一事务，任一步失败整体回滚。
      parameters: [{ $ref: '#/components/parameters/TopicId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UnitAllocationBatch' } } } }
      responses: { '204': { description: 分配已确认并生效 }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationError' } }
  '/topics/{topicId}/unit-allocations:confirm-plan':
    put:
      tags: [Indicators]
      operationId: confirmUnitAllocationPlan
      description: 当前有效牵头一次提交全部启用时间阶段的完整单位分配方案；任一阶段缺失、任一单位指标未覆盖或阶段合计不等于课题本阶段目标时拒绝，全部阶段在同一事务中生效。
      parameters: [{ $ref: '#/components/parameters/TopicId' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UnitAllocationPlanBatch' } } } }
      responses: { '204': { description: 全部分配方案已确认并生效 }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationError' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Size: { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 20 } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 100 } }
    UserId: { name: userId, in: path, required: true, schema: { type: string } }
    RoleId: { name: roleId, in: path, required: true, schema: { type: string } }
...
    TopicId: { name: topicId, in: path, required: true, schema: { type: string } }
...
  responses:
    Unauthorized: { description: 未登录或令牌失效, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Forbidden: { description: 没有功能或数据权限, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Conflict: { description: 唯一约束、状态或乐观锁冲突, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    ValidationError: { description: 业务或字段校验失败, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, instance, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, example: TOPIC_NOT_OPERATIONAL }
        traceId: { type: string }
        fieldErrors: { type: array, items: { type: object, properties: { field: { type: string }, message: { type: string } } } }
...
    User:
      type: object
      required: [id, username, name, roleId, roleName, enabled]
      properties:
        id: { type: string }
        username: { type: string }
        name: { type: string, description: 单位联系人姓名 }
        unitId: { type: [string, 'null'] }
        roleId: { type: string }
        roleName: { type: string }
        phone: { type: [string, 'null'] }
        email: { type: [string, 'null'], format: email }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
    UserCreateRequest:
      type: object
      required: [username, roleId, name, password]
      description: 课题单位账号的单位由后端按用户名自动创建，内外属性由角色确定。
      properties: { username: { type: string }, roleId: { type: string }, name: { type: string }, phone: { type: string }, email: { type: string, format: email }, enabled: { type: boolean, default: true }, password: { type: string, minLength: 8, maxLength: 72, writeOnly: true } }
    UserCreateResponse:
      type: object
      required: [user]
      properties: { user: { $ref: '#/components/schemas/User' } }
    UserUpdateRequest:
      type: object
      properties: { username: { type: string }, roleId: { type: string }, name: { type: string }, phone: { type: string }, email: { type: string, format: email } }
    PasswordChangeRequest:
      type: object
      required: [password]
      properties: { password: { type: string, minLength: 8, maxLength: 72, writeOnly: true } }
    UserPage:
      type: object
      required: [items, page, size, total]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/User' } }, page: { type: integer }, size: { type: integer }, total: { type: integer, format: int64 } }
    Role:
      type: object
      required: [id, code, name, pagePermissions, actionPermissions, enabled, builtIn]
      properties: { id: { type: string }, code: { type: string }, name: { type: string }, description: { type: string }, pagePermissions: { type: array, items: { type: string } }, actionPermissions: { type: array, items: { type: string } }, enabled: { type: boolean }, builtIn: { type: boolean } }
    RolePermissionRequest:
      type: object
      required: [pagePermissions, actionPermissions]
      properties: { pagePermissions: { type: array, uniqueItems: true, items: { type: string } }, actionPermissions: { type: array, uniqueItems: true, items: { type: string } }, enabled: { type: boolean } }
...
    Unit:
      type: object
      required: [id, code, name, internal, enabled]
      properties: { id: { type: string }, code: { type: string }, name: { type: string }, internal: { type: boolean }, enabled: { type: boolean } }
    TopicStatus: { type: string, enum: [DRAFT, ACTIVE, PAUSED, CLOSED] }
    Topic:
      type: object
      required: [id, code, name, leadUnitId, status, enabled, recordVersion]
      properties: { id: { type: string }, code: { type: string }, name: { type: string }, summary: { type: string }, leadUnitId: { type: string }, status: { $ref: '#/components/schemas/TopicStatus' }, enabled: { type: boolean }, startDate: { type: string, format: date }, endDate: { type: string, format: date }, recordVersion: { type: integer }, members: { type: array, items: { $ref: '#/components/schemas/TopicMembership' } } }
    TopicWriteRequest:
      type: object
      required: [code, name, leadUnitId]
      properties: { code: { type: string }, name: { type: string }, summary: { type: string }, leadUnitId: { type: string }, participantUnitIds: { type: array, items: { type: string } }, startDate: { type: string, format: date }, endDate: { type: string, format: date }, recordVersion: { type: integer } }
...
    TopicMembership:
      type: object
      required: [id, topicId, unitId, membershipType, enabled]
      properties: { id: { type: string }, topicId: { type: string }, unitId: { type: string }, unitName: { type: string }, membershipType: { type: string, enum: [LEAD, PARTICIPANT] }, enabled: { type: boolean } }
...
    UnitIndicatorAllocation:
      type: object
      required: [id, topicId, unitId, nodeId, indicatorDefinitionId, targetQuantity, status, version]
      properties: { id: { type: string }, topicId: { type: string }, unitId: { type: string }, nodeId: { type: string }, indicatorDefinitionId: { type: string }, targetQuantity: { type: integer, minimum: 0 }, status: { type: string, enum: [DRAFT, PUBLISHED] }, version: { type: integer } }
    UnitAllocationBatch:
      type: object
      required: [nodeId, allocations]
      properties: { nodeId: { type: string }, draftVersion: { type: integer, minimum: 0, maximum: 2147483647 }, allocations: { type: array, maxItems: 5000, items: { type: object, required: [unitId, indicatorDefinitionId, targetQuantity], properties: { unitId: { type: string }, indicatorDefinitionId: { type: string }, targetQuantity: { type: integer, minimum: 0, maximum: 2147483647 } } } } }
    UnitAllocationPlanBatch:
      type: object
      required: [stages]
      properties: { stages: { type: array, minItems: 1, maxItems: 100, items: { $ref: '#/components/schemas/UnitAllocationBatch' } } }
...